Move backupDownSampler into util and add tests
